A new python notebook
====================
HR dataset
=====================
Google Advanced Data Analytics Capstone project: Providing data-driven suggestions for HR


Introduction

Description and deliverables
This capstone project is an opportunity for me to analyze a dataset and build predictive models that can provide insights to the Human Resources (HR) department of a large consulting firm.

Upon completion, I will have two artifacts that would be presented to my future employers, a brief one-page summary of this project that would presented to external stakeholders as the data professional in Salifort Motors. To achieve this given project question, I would be using either a regression model or machine learning model to predict whether or not an employee will leave the company. 

In my deliverables, i will include the model evaluation (and interpretation ), data visualization(s) that is directly related to the question asked, ethical considerations, and the resources i used to troubleshoot and find answers or solutions.